Created a base class for initialisation functionality. Refactored GradientDescent. Added gitignore
