+++ .gitignore
@@ -1,0 +1,1 @@
+minimise_functions_gradient_descent_and_normal_equation/__pycache__/
