+++ minimise_functions_gradient_descent_and_normal_equation/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+description-file = README.md
